requirements.txt
psycopg2-binary==2.7.5
flask==1.1.2
passlib==1.7.2
Flask_SQLAlchemy==2.3.2
gunicorn==20.0.4